% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMultLognGPD.R
\name{dMultLognGPD}
\alias{dMultLognGPD}
\title{Density of a multivariate lognormal-GPD mixture, where the GPD has a
Gumbel dependence structure}
\usage{
dMultLognGPD(x, p, mu, Psi, gpdpars, gammap)
}
\arguments{
\item{x}{(dx1) positive vector: points where the function is evaluated.}

\item{p}{real, 0<p<1: prior probability}

\item{mu}{(dx1) vector: log-mean of the multivariate lognormal distribution.}

\item{Psi}{positive definite symmetric (dxd) matrix: covariance matrix of
the underlying multivariate normal distribution.}

\item{gammap}{real > 1: parameter of the Gumbel copula.}

\item{gdppars}{(2xd)-vector of positive real numbers: matrix whose i-th
column contains the scale and shape parameters of the i-th
marginal generalized Pareto distribution.}
}
\value{
dmix n-vector: n values of the
d-variate lognormal - generalized Pareto mixture density.
}
\description{
This function evaluaates the density of a multivariate lognormal-GPD
mixture.
}
\examples{
y <- dMultLognGPD(c(1,2),.5,c(0,0),diag(2),rbind(c(1,1.5),c(.25,5)),2)

}
