% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weiCopLik.R
\name{weiCopLik}
\alias{weiCopLik}
\title{Weighted Gumbel copula log-likelihood}
\usage{
weiCopLik(x, y, p_weights, gpdpars)
}
\arguments{
\item{x}{scalar > 1: values of the Gumbel copula parameter.}

\item{y}{data matrix (nxd): observed data.}

\item{p_weights}{numerical vector (nx1) with elements in (0,1): weights
of the observations (in the EM algorithm, posterior probabilities).}

\item{gdp_pars}{matrix (2xd) with non-negative elements: scale (1st
row) and shape (2nd row) parameters of the marginal GPDs.}
}
\value{
llik real: numerical value of the weighted log-likelihood
function
}
\description{
This function evaluates the Gumbel copula log-likelihood
function computed with weighted observations.
}
\examples{
ysim <- rMultLognGPD(100,.9,2,c(0,0),diag(2),c(.25,.5),c(1,1.5),2)
x0 <- c(.7,.2,1.3,.8,1.7)
res <- EMlogngpdmix(x0, y, 1000)
llik <- weiGpdLik(c(res$beta,res$xi),y,res$post)

}
