% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weiGpdLik.R
\name{weiGpdLik}
\alias{weiGpdLik}
\title{Weighted GPD log-likelihood}
\usage{
weiGpdLik(x, type.cop, y, p_weights)
}
\arguments{
\item{x}{numerical vector (2dx1): values of the parameters \eqn{\beta}
and \eqn{\xi} for each marginal: (\eqn{\beta_1},\eqn{\xi_1},\eqn{\beta_2},\eqn{\xi_2},...).}

\item{y}{data matrix (nxd): observed data.}

\item{p_weights}{numerical vector (nx1) with elements in (0,1): weights
of the observations (in the EM algorithm, posterior probabilities).}
}
\value{
llik real: numerical value of the weighted GPD log-likelihood
function
}
\description{
This function evaluates the zero-mean generalized Pareto log-likelihood
function computed with weighted observations.
}
\examples{
ysim <- rMultLognGPD(100,.9,2,c(0,0),diag(2),c(.25,.5),c(1,1.5),2)
x0 <- c(.7,.2,1.3,.8,1.7)
res <- EMlogngpdmix(x0, y, 1000)
llik <- weiGpdLik(c(res$beta,res$xi),y,res$post)

}
