% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMult.R
\name{EMMult}
\alias{EMMult}
\title{Multivariate estimation}
\usage{
EMMult(ymix, eps, maxiter)
}
\arguments{
\item{ymix}{data matrix (nxd): observed data.}

\item{eps}{positive real: tolerance in the stopping criterion of the EM algorithm.}

\item{maxiter}{positive integer: maximum number of iterations of the EM algorithm.}

\item{nboot}{positive integer: number of bootstrap replications for the
computation of the standard errors (defaults to 0).}
}
\value{
A list with the following elements is returned:
"pars" = estimated values of the parameters,
"loglik" = maximimzed log-likelihood,
"niter" = number of iterations,
"post_p" = posterior probabilities of all observations,
bootEst = matrix of parameter estimates at each bootstrap replications (only if nboot > 0).
bootStd = bootstrap standard errors of each parameter (only if nboot > 0).
}
\description{
This function estimates a multivariate lognormal - generalized Pareto mixture
by means of the EM algorithm. Optionally, bootstrap standard errors are
computed via parallel computing.
}
\examples{
y <- rMultLognGPD(100,.9,2,c(0,0),diag(2),c(.25,.5),c(1,1.5),2)
x0 <- c(.7,.2,1.3,.8,1.7)
res <- EMlogngpdmix(x0, y, 1000)

}
