% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMultLognGPD.R
\name{rMultLognGPD}
\alias{rMultLognGPD}
\title{Simulation of a multivariate lognormal-GPD mixture, where the GPD has a
Gumbel dependence structure}
\usage{
rMultLognGPD(n, p, d, mu, Psi, gpdpars, gammap)
}
\arguments{
\item{n}{positive integer: number of observations sampled.}

\item{p}{real, 0<p<1: prior probability.}

\item{d}{positive integer: dimension of the distribution.}

\item{mu}{(dx1) vector: log-mean of the multivariate lognormal distribution.}

\item{Psi}{positive definite symmetric (dxd) matrix: covariance matrix of
the underlying multivariate normal distribution.}

\item{gdppars}{(2xd)-matrix of positive real numbers: matrix whose i-th
column contains the scale and shape parameters of the i-th
marginal generalized Pareto distribution.}
}
\value{
ysim (n x d) matrix: n random vectors from the
d-variate lognormal - generalized Pareto mixture.
}
\description{
This function simulates a multivariate lognormal-GPD mixture.
}
\examples{
gpdpars <- cbind(c(1,0.5),c(1.5,0.25))
ysim <- rMultLognGPD(100,.9,2,c(0,0),diag(2),gpdpars,2)

}
